requirements.txt
fastapi
uvicorn
gspread
oauth2client
polars
langchain
pandas
openai
langchain_openai
langchain-google-genai
langchain-mcp-adapters
mcp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
unstructured
langchain-googledrive           
langgraph
langchain[anthropic]
